RAG_App/requirements.txt
gradio
faiss-cpu
sentence-transformers
torch
pickle-mixin
